backend/requirements.txt
fastapi
uvicorn
pydoll-python
sqlalchemy
httpx
pandas